+++ CHANGELOG.md
@@ -4,5 +4,6 @@ Changelog
 v0.1.0 (not released)
 ---------------------
 
+- Class edition (name, extends, implements)
 - Use statements edition
 - Namespace edition